Fix argparse type

+++ code/algorithm/main.py
@@ -10,13 +10,13 @@ from pathlib import Path
 from typing import List
 
 import numpy as np
-import wandb
 from datasets import load_data
 from factories import ModelFactory, PreprocessorFactory
 from loggers import JSONLLogger, StreamLogger, WandbLogger
 from termcolor import colored
 from utilities import rescale
 
+import wandb
 from config import Dataset, Model, Noise
 
 
@@ -163,7 +163,7 @@ def parse_args(args: List[str]) -> argparse.Namespace:
     )
     data_parser.add_argument(
         "--subset",
-        type=int,
+        type=float,
         default=0.9,
         help="A float between 0 and 1, the amount of training data to train on.",
     )
